Squashed 'mpfs_hal/' changes from 9ce11f6..2a91d75
2a91d75 Added more debug information to ddr
022e9ab Minor improvements after refactor
f8abcad Added missing call from last commit
2d8aa1f     Added assert() to check lim size
da24333 SAR 125332 add GPIO_INTERRUPT_FAB_CR support
584f54f Updated version to 1.8.143
c79867a Some tidy-up
5db8a4d Minor code clean-up
3341d8d Astetic code change to make clearer
1575662 Merge branch 'slight-code-refactor' into revc-ddr-updates
7a93c62 Updated version to v1.8.142
4ba63cf Added weak ddr progress function
0a182ba warning removal
0e93a28 Version update to 1.8.142
e35b4ed Merge branch 'slight-code-refactor' into revc-ddr-updates
ece4154 Updated the copyright notice year ending to 2022
024eeca Removed some warnings when compiled with HSS
a46d1b6 Merge branch 'hss_error_removal' into revc-ddr-updates
a63dacd ddr update v0.4.013
ad47b38 ddr update v0.4.012 default settings update for LPDDR4
0aa1661 ddr update v0.4.011    ADD_CMD_CLK_MOVE_ORDER changed from 0,1,2 to 1,2,0
9993648 ddr update v0.4.010    LIBERO_SETTING_RPC_156_VALUE default changed from 1 to 6
7eb78ac ddr update v0.4.009    vrgen, modify during write leveling for DDR3 corrected
98fe9d1 ddr update v0.4.008     DQ/DQS push order has been parameterised rder has been parameterised
8384a65 ddr update v0.4.007 Corrected write_latency print message
fab36f2 DDR update - traing time improvement
2e5755e Refactored delay() routine to use mcycle and MSS clk from MSS Configurator
0cc8ca5 ddr update v0.4.005 When LIBERO_FAST_START, now slects random pattern
6aa42d3 ddr update v0.4.004    Updated tip_register_status() to show dual ranks
a6612db Merge branch 'correct-legacy-interrupt-naming' into revc-ddr-updates
fd17f8e ddr update v0.4.003  Added FAST_START option
c0c8d63 ddr update v0.4.001     |  Fixed DDR3 DDR_1333_MHZ define to match Libero gen version
df48b74 ddr update v0.3.030  Added setting of rpc136, required for board tuning to pass
d86a29a ddr update v0.3.029   LIBERO_SETTING_REFCLK_DDR3_1067_NUM_OFFSETS changed 2 to 1
872c83b ddr update v0.3.028  ddr3_address_cmd_training() routine added
823f3f3 ddr update v0.3.027    ddr3 mod- vrgen, modify during write leveling
5eba480 ddr update 0.3.026  SW_TRAING_BCLK_SCLK_OFFSET changed from 0 to 5
e3dbe8b ddr updated v0.3.025 LPDDR4@1600 ref clk offsets 4,3,2,4 changed to 3,4,2,5
96ebe80 ddr update v0.3.024  lpddr4_manual_training() improved
29fc72d ddr update v0.3.023  Changing the common mode of the Receiver to low common mode
f693521 DDR update v0.3.022   DDR_VERIFY_PATTERN_IN_CACHE added tests
a391632 ddr update  v0.3.021   Turn off ODT during write leveling
6114d1a ddr update v0.3.020  retrain reset for better performance
6000e5c Merge branch 'bugfix/sar120825-resolve-naming-mismatch' into revc-ddr-updates
4a8a251 SAR122487 - LPDDR4 Firmware DQDQS Alignment margin filter is set on Status1 instead of Status2
547333e SAR121393 relates to DDR3 robustness when ECC not being used
cb3afc3 DDR update v0.3.017
fb91d9e Merge branch 'develop' into revc-ddr-updates
db5fb96 DDR updated - Updated DDR3, LPDDR3, and DDR4, changes related to REV C testing
44d84ed Updated warm reset to turn-off, then on at start
65957de Update to print details of second rank on debug output

git-subtree-dir: mpfs_hal
git-subtree-split: 2a91d758caac3208efb7464f0cdebe411b6abbd9

+++ common/mss_util.c
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2019-2021 Microchip FPGA Embedded Systems Solutions.
+ * Copyright 2019-2022 Microchip FPGA Embedded Systems Solutions.
  *
  * SPDX-License-Identifier: MIT
  *
@@ -77,7 +77,7 @@ void __enable_local_irq(uint8_t local_interrupt)
     ASSERT(local_interrupt > (int8_t)0);
     ASSERT( (local_interrupt <= LOCAL_INT_MAX));
 
-    uint8_t mhart_id = read_csr(mhartid);
+    uint8_t mhart_id = (uint8_t)read_csr(mhartid);
 
     if((local_interrupt > (int8_t)0) && (local_interrupt <= LOCAL_INT_MAX))
     {
